Add SelectMetabolite which picks next metabolite to remove.
SelectMetabolite.{h|cu} adds both CPU code and a kernel
to compute the input and output counts for each
metabolite and then select the best one (see design doc).

Also add code to the setup file to initialize counts
and balancedMetabolites array.

TODO: the {h|d}_{input|output}Count arrays are declared and
in allocated in EFMGenerator but this should be done in Setup.